api/requirements.txt
Django==2.1.3
djangorestframework==3.9.0
pyyaml==3.13
djangorestframework-jwt==1.11.0
ipython==7.2.0
